feat(standalone-canary-analysis): add SCAPE complete event listener and archive the status response object in the object store. (#593)
