--- kayenta-standalone-canary-analysis/src/main/java/com/netflix/kayenta/canaryanalysis/config/StandaloneCanaryAnalysisModuleConfiguration.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2018 Nike, inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License")
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.kayenta.canaryanalysis.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConditionalOnProperty("kayenta.standalone-canary-analysis.enabled")
-@ComponentScan({"com.netflix.kayenta.canaryanalysis"})
-@Slf4j
-public class StandaloneCanaryAnalysisModuleConfiguration {}
