+++ kayenta-standalone-canary-analysis/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.netflix.kayenta.canaryanalysis.config.StandaloneCanaryAnalysisModuleConfiguration
+com.netflix.kayenta.standalonecanaryanalysis.config.StandaloneCanaryAnalysisModuleConfiguration
